Move components to reflect table of contents
