--- src/ContactMe/ContactMe.js
@@ -1,10 +1,0 @@
-function ContactMe() {
-    return (
-      <div className="contact-me-container">
-        <h1>Contact Me</h1>
-      </div>
-    );
-  }
-  
-  export default ContactMe;
-
