+++ src/ContentSections/Home/ContactMe/ContactMe.js
@@ -1,0 +1,9 @@
+function ContactMe() {
+  return (
+    <div className="contact-me-container">
+      <h1>Contact Me</h1>
+    </div>
+  );
+}
+
+export default ContactMe;
