+++ src/common/WritingTreeHeader/WritingTreeHeader.js
